source/requirements.txt
alembic==1.4.2
aioredis==1.3.1
aiosmtplib==1.1.3
blessings==1.7
gino==0.8.6
marshmallow==3.5.1
marshmallow-enum==1.5.1
marshmallow-jsonapi==0.23.1
openpyxl==3.0.3
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytest-mock==3.0.0
pytest-sanic==1.6.0
python-dotenv==0.12.0
pytz==2019.3
sanic==19.12.2
# sanic-openapi==0.6.1
uptime==3.0.1
Werkzeug==1.0.1

# see for details https://github.com/huge-success/sanic-openapi/pull/157 and 168
git+git://github.com/bolshoydi/sanic-openapi.git@4567e81

git+ssh://git@github.com/bistegra/fwork.git@0.0.2#egg=fwork_common
